Maintenance compile-time calculation of factorial

+++ savchuk.aleksei/B3/tools.hpp
@@ -8,8 +8,8 @@ namespace savchuk
 {
   constexpr unsigned long long calcucateFactorial(unsigned int n)
   {
-    size_t res = 1;
-    for (size_t i = n; i > 0; --i)
+    unsigned long long res = 1;
+    for (unsigned int i = n; i > 0; --i)
     {
       res *= i;
     }
